+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Throw an error when trying to set a Document's data record to an invalid value
 * Fix getting the path of fields in a subschema
 * Add `Schema#getFieldChain(path)` and `Field#getFieldChain()`
+* Add `FieldValue` class and use it for the new `Field#getDocumentValues(document)` method
 
 ## 1.0.8 (2019-01-12)